+++ javax/src/main/java/me/panpf/javax/io/Filex.java
@@ -1238,7 +1238,7 @@ public class Filex {
      * @param blockSize size of a block, replaced by 512 if it's less, 4096 by default.
      */
     public static void forEachBlock(@NotNull File file, int blockSize, @NotNull Action2<byte[], Integer> action) throws IOException {
-        byte[] arr = new byte[Numberx.coerceAtLeast(blockSize, IOStreamx.MINIMUM_BLOCK_SIZE)];
+        byte[] arr = new byte[Rangex.coerceAtLeast(blockSize, IOStreamx.MINIMUM_BLOCK_SIZE)];
 
         InputStream input = inputStream(file);
         try {