+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * :sparkles: Add Mathx
 * :fire: Numberx scale method move to Mathx
 * :fire: Numberx in, notIn method move to Rangex
+* :fire: Numberx coerceAtLeast, coerceAtMost method move to Rangex
 
 ## 0.6.6